refactor: moved parsers and interfaces to new separate files for improved readability
